refac: change num_elements to usize
Clean up some variable definitions

+++ src/config.rs
@@ -88,7 +88,7 @@ pub struct LineSegment {
     #[pyo3(get)]
     pub length: f64,
     #[pyo3(get)]
-    pub num_elements: i32,
+    pub num_elements: usize,
 }
 
 impl LineSegment {
